+++ resources/views/admin/dashboard/index.php
@@ -11,6 +11,7 @@ use Hirtz\Skeleton\Modules\Admin\Widgets\DirectoryAlert;
 use Hirtz\Skeleton\Modules\Admin\Widgets\EnvironmentAlert;
 use Hirtz\Skeleton\Modules\Admin\Widgets\MigrationAlert;
 use Hirtz\Skeleton\Modules\Admin\Widgets\Navs\DashboardHeader;
+use Hirtz\Skeleton\Modules\Admin\Widgets\SentryAlert;
 use Hirtz\Skeleton\Web\View;
 use Hirtz\Skeleton\Widgets\Panels\Dashboard;
 
@@ -20,4 +21,5 @@ echo DashboardHeader::make();
 echo MigrationAlert::make();
 echo DirectoryAlert::make();
 echo EnvironmentAlert::make();
+echo SentryAlert::make();
 echo Dashboard::make();